requirements.txt
openai==0.27.2
python-dotenv==1.0.0
docker

# OpenAI and Generic plugins import
openapi-python-client==0.13.4
